consolidated to types, collections and operators
